[maglev][arm] Use enum for Conditions in constants-arm.h
Maglev uses duck typing for Condition, but it expects an enum.

Bug: v8:7700
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4567554
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#87945}

+++ src/codegen/x64/assembler-x64.h
@@ -63,7 +63,7 @@ class MaglevSafepointTableBuilder;
 
 // Utility functions
 
-enum Condition : uint8_t {
+enum Condition : int {
   overflow = 0,
   no_overflow = 1,
   below = 2,
